requirements_dev.txt
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
black==21.5b2
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
coverage==5.5
decorator==5.0.9
defusedxml==0.7.1
docutils==0.17.1
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.17
idna==2.10
importlib-metadata==4.5.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==23.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
mpmath==1.2.1
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pipdeptree==2.0.0
pkginfo==1.7.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySimpleGUI==4.43.0
python-dateutil==2.8.1
pytoxo2==0.1.dev1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
readme-renderer==29.0
regex==2021.4.4
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
Send2Trash==1.5.0
six==1.16.0
smmap==4.0.0
sympy==1.8
tabulate==0.8.9
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
twine==3.4.1
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.1
